remove binary endpoint and expose some client behaviours in internal namespace
* initialize endpoint names per defatul
