+++ src/OpenRiaServices.DomainServices.Hosting/Framework/Services/Behaviors/SoapQueryBehavior.cs
@@ -1,14 +1,12 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using System.ServiceModel;
-using OpenRiaServices;
 using System.ServiceModel.Channels;
 using System.ServiceModel.Description;
 using System.ServiceModel.Dispatcher;
-using OpenRiaServices.DomainServices.Hosting;
 using System.Xml;
 
-namespace OpenRiaServices.DomainServices.Hosting
+namespace OpenRiaServices.DomainServices.Hosting.Behaviors
 {
     /// <summary>
     /// A SOAP endpoint behavior which injects a message inspector that parses query headers.
